Move final Markdown rendering pass to a submodule
